整理Sentence的单元测试
